// teamplate/zaui-mmenu/src/@types/mmenu.d.ts
export interface LocalizationName {
  language: Language
  name?: string
}

export enum Language {
  En = 'en',
  Ja = 'ja',
  Ko = 'ko',
  Th = 'th',
  Vi = 'vi',
  Zh = 'zh',
}
// ...
export enum Status {
  Available = 'available',
  Unavailable = 'unavailable',
}

export interface Topping {
  toppingId: string
  toppingName: string
  toppingItems?: ToppingItem[]
  isRequired: boolean
  isMultipleSecleted: boolean
  restaurantId: string
  toppingOrdinalNumber: number
  toppingType?: ToppingType
  limitQuantity?: number
  hasQuantity?: boolean
  attachedDishes?: ToppingAttachedDish[]
  localizationNames?: LocalizationName[]
}

export interface ToppingAttachedDish {
  attachedDishName: string
  attachedDishId: string
  price: number
  ordinalNumber?: number
  image?: string
}

export interface ToppingItem {
  name: string
  price: number
  isDefault: boolean
  toppingItemId: string
  status: Status
  quantity: number
  localizationNames?: LocalizationName[]
}

export enum ToppingType {
  AttachedDish = 'attachedDish',
  DishGroup = 'dishGroup',
  Topping = 'topping',
}